# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Your API description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Регистрация"
                ],
                "summary": "Регистрация",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "name",
                                    "address",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+79123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "ул. Пушкина, д. 1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь зарегистрирован"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+79123456789"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "История заказов",
                "description": "Получение списка заказов. Требуется авторизация.",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvYXBpL2xvZ2luIiwiaWF0IjoxNjMyNTQ5MjAwLCJleHAiOjE2MzI1NTI4MDAsIm5iZiI6MTYzMjU0OTIwMCwianRpIjoiMjNkOGFiYzEyM2QxMmFiYyIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.HsQtB1bqkYYdxV9nqjqG9zJpWBmH4nqzxH3mJfU1E8Q"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cписок заказов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "total_price": {
                                                "type": "number",
                                                "example": 2500
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            },
                                            "comment": {
                                                "type": "string",
                                                "example": "Доставить до 18:00"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-19T16:53:42.000000Z"
                                            },
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "product_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "quantity": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "price": {
                                                            "type": "number",
                                                            "example": 1250
                                                        },
                                                        "product": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Говядина"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "Говядина"
                                                                },
                                                                "price": {
                                                                    "type": "number",
                                                                    "example": 1250
                                                                },
                                                                "category": {
                                                                    "type": "string",
                                                                    "example": "Мясо"
                                                                },
                                                                "in_stock": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нет доступа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Оформить заказ",
                "description": "Создание заказа. Требуется авторизация.",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT токен в формате: Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Доставить до 18:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован или неверный токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Товары"
                ],
                "summary": "Список товаров",
                "description": "Авторизация не требуется",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Говядина"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Говядина"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 500
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Мясо"
                                            },
                                            "in_stock": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT токен авторизации",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Регистрация",
            "description": "Регистрация"
        },
        {
            "name": "Авторизация",
            "description": "Авторизация"
        },
        {
            "name": "Заказы",
            "description": "Заказы"
        },
        {
            "name": "Товары",
            "description": "Товары"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}